requirements.txt
flask
flask_restful
flask_cors
flask_sqlalchemy

MySQLdb
werkzeug

google-auth>=1.27.0
google_auth_oauthlib>=0.4.2
oauthlib>=3.1.0pathlib>=1.0.1vendor>=0.1
requests>=2.25.1
requests-oauthlib>=1.3.0
